allow missing defs
Looks like sometimes this is a valid thing. I notice in particular,
implementations of Iterator have two child items, one for the `next()`
method, and another for the Item type, which sometimes doesn't have a
corresponding def. I feel like this shouldn't happen, but if nothing can
be done about it, allow callers to ignore it.